requirements.txt
APScheduler==3.7.0
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.5
coreapi==2.3.3
coreschema==0.0.4
coverage==5.5
cryptography==3.4.8
defusedxml==0.7.1
Deprecated==1.2.13
Django==3.2.7
django-apscheduler==0.6.0
django-cors-headers==3.8.0
django-filter==2.4.0
django-guardian==2.4.0
django-oauth-toolkit==1.5.0
django-request-logging==0.7.3
djangorestframework==3.12.4
djangorestframework-csv==2.1.1
djangorestframework-xml==2.0.0
djangorestframework-yaml==2.0.0
drf-renderer-xlsx==0.4.1
drf-yasg==1.20.0
drfdocs==0.0.11
et-xmlfile==1.1.0
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
jwcrypto==1.0
Markdown==3.3.4
MarkupSafe==2.0.1
oauthlib==3.1.1
openpyxl==3.0.8
packaging==21.0
pipdeptree==2.1.0
psycopg2==2.9.1
pycparser==2.20
Pygments==2.10.0
pyparsing==2.4.7
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
tzlocal==2.1
unicodecsv==0.14.1
uritemplate==3.0.1
urllib3==1.26.6
wrapt==1.12.1
